requirements.txt
streamlit
psycopg2
sqlalchemy
pandas
requests
python-dotenv
beautifulsoup4